python/root/debian.py: /etc/X11/Xsession.options might not exist

+++ python/root/debian.py
@@ -71,11 +71,11 @@ def pi_setup(user_name):
                            '\n}\n')
         docker_json.write_text(docker_json_txt, encoding='utf-8')
 
-    x11_opts = Path('/etc/X11/Xsession.options')
-    x11_opts_txt = x11_opts.read_text(encoding='utf-8')
-    if re.search('^use-ssh-agent$', x11_opts_txt, flags=re.M):
-        x11_opts_txt = re.sub('use-ssh-agent', '# use-ssh-agent', x11_opts_txt)
-        x11_opts.write_text(x11_opts_txt, encoding='utf-8')
+    if (x11_opts := Path('/etc/X11/Xsession.options')).exists():
+        x11_opts_txt = x11_opts.read_text(encoding='utf-8')
+        if re.search('^use-ssh-agent$', x11_opts_txt, flags=re.M):
+            x11_opts_txt = re.sub('use-ssh-agent', '# use-ssh-agent', x11_opts_txt)
+            x11_opts.write_text(x11_opts_txt, encoding='utf-8')
 
 
 def prechecks():
